Add Levenshtein algorithm to all dictionaries.
