+++ src/main/java/Telia/EstonianEnglishDictionary/Service/EnglishWordService.java
@@ -1,5 +1,6 @@
 package Telia.EstonianEnglishDictionary.Service;
 
+import Telia.EstonianEnglishDictionary.Algorithm.LevenshteinAlgorithm;
 import Telia.EstonianEnglishDictionary.Model.EnglishWord;
 import Telia.EstonianEnglishDictionary.Model.Translation;
 import Telia.EstonianEnglishDictionary.Repository.EnglishWordsRepository;
@@ -7,14 +8,13 @@ import Telia.EstonianEnglishDictionary.Repository.TranslationRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.Arrays;
-import java.util.LinkedList;
-import java.util.List;
-import java.util.Optional;
+import java.util.*;
 import java.util.stream.Collectors;
 
 @Service
 public class EnglishWordService {
+
+    private static final LevenshteinAlgorithm levenshteinAlgorithm = new LevenshteinAlgorithm();
 
     @Autowired
     private TranslationRepository translationRepository;
@@ -47,45 +47,25 @@ public class EnglishWordService {
         return enWord;
     }
 
-    public static int costOfSubstitution(char a, char b) {
-        return a == b ? 0 : 1;
-    }
-
-    public static int min(int... numbers) {
-        return Arrays.stream(numbers)
-                .min().orElse(Integer.MAX_VALUE);
-    }
-
-    public int calculateLevenshtein(String x, String y) {
-        int[][] dp = new int[x.length() + 1][y.length() + 1];
-
-        for (int i = 0; i <= x.length(); i++) {
-            for (int j = 0; j <= y.length(); j++) {
-                if (i == 0) {
-                    dp[i][j] = j;
-                }
-                else if (j == 0) {
-                    dp[i][j] = i;
-                }
-                else {
-                    dp[i][j] = min(dp[i - 1][j - 1] + costOfSubstitution(x.charAt(i - 1), y.charAt(j - 1)),
-                                            dp[i - 1][j] + 1,
-                                            dp[i][j - 1] + 1);
-                }
-            }
-        }
-
-        return dp[x.length()][y.length()];
-    }
-
-    public List<EnglishWord> translate(String word) {
-        List<EnglishWord> similarWords =  englishWordsRepository.findAll()
+    public Map<String, Object> translate(String word) {
+        boolean foundExact = true;
+        List<EnglishWord> suitableWords =  englishWordsRepository.findAll()
                 .stream()
-                .filter(x -> calculateLevenshtein(word, x.getWord()) <= 2)
+                .filter(x -> x.getWord().equals(word))
                 .collect(Collectors.toList());
+        if (suitableWords.size() == 0) {
+            foundExact = false;
+            suitableWords =  englishWordsRepository.findAll()
+                    .stream()
+                    .filter(x -> levenshteinAlgorithm.calculateLevenshtein(word, x.getWord()) <= 2)
+                    .collect(Collectors.toList());
+        }
         //List<Translation> translations = new LinkedList<>();
         //similarWords.forEach(w -> translations.addAll(w.getEquivalents()));
-        return similarWords;
+        Map<String, Object> words = new LinkedHashMap<>();
+        words.put("wordList", suitableWords);
+        words.put("foundExact", foundExact);
+        return words;
     }
 
     public void deleteWordAndTranslations(Long id) {
